This change removes dshistory and add History.js to Spiffy UI.
History.js has support for HTML5 which works much better on Chrome and Opera.  It also has better support for URL variables which allows us to have much smaller URLs.  We still have the problem of the full host name showing up as an URL variable in HTML4 browsers.  We are working with the developer to make that work.

This change also remove the navigation cookie from SpiffyUI.org.  The navigation cookie was in place to make send you back to the page you were on when you refresh the browser.  That works fine for refreshing, but it sends you to an unexpected page if you were looking at a specific page, close that page, and then go back to the root of the application.  You asked for the root, but the cookie remembers that you used to be at a different page.

The long story short here is that the cookie isn't necessary anymore because we have bookmarkable URLs.

We have also updated the sitemap.xml file generation to use our new URLs.  We no longer have the #! syntax in our standard URLs, but we still support them for the sitemap so we can take advantage of AJAX crawling.

This change also adds the new debugging video to the dev mode page.