Make most Restility methods protected->public
We want to support extensiblity with the authentication
mechanism, but currently Restility only exposes the
necessary methods for extention to objects within the
.rest package.  While its fairly cheesy to just
open up the class for public access, it really needs
a fairly major redesign in order to do this "properly".
Therefore, the short term solution is just just permit
access and we can work on the redesign over the next few
weeks.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>