We are no longer disabling the MVSB control.
When you type a value into the MVSB control we disable it for a moment while we are updating the values.  This means the text field can't take focus.  If the user presses the backspace key in that small time window it will cause the browser to go back in history instead of deleting the previous character.  This change stops disabling the control since we handle value updates at this time well and we want to consume all of the key events in the text field.